Move method to common class.
