+++ hadoop-ozone/recon/src/main/java/org/apache/hadoop/ozone/recon/ReconControllerModule.java
@@ -19,8 +19,6 @@ package org.apache.hadoop.ozone.recon;
 
 import java.io.IOException;
 import java.lang.reflect.Constructor;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 
 import org.apache.hadoop.hdds.conf.OzoneConfiguration;
@@ -52,15 +50,13 @@ import org.apache.hadoop.ozone.recon.tasks.ReconTaskController;
 import org.apache.hadoop.ozone.recon.tasks.ReconTaskControllerImpl;
 import org.apache.hadoop.security.UserGroupInformation;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
 import com.google.inject.multibindings.Multibinder;
 import static org.apache.hadoop.hdds.scm.cli.ContainerOperationClient.newContainerRpcClient;
-import static org.apache.hadoop.ozone.om.OMConfigKeys.OZONE_OM_INTERNAL_SERVICE_ID;
-import static org.apache.hadoop.ozone.om.OMConfigKeys.OZONE_OM_SERVICE_IDS_KEY;
+import static org.apache.hadoop.ozone.OmUtils.getOzoneManagerServiceId;
 
 import org.apache.ratis.protocol.ClientId;
 import org.hadoop.ozone.recon.codegen.ReconSqlDbConfig;
@@ -166,32 +162,6 @@ public class ReconControllerModule extends AbstractModule {
       LOG.error("Error in provisioning OzoneManagerProtocol ", ioEx);
     }
     return ozoneManagerClient;
-  }
-
-  @VisibleForTesting
-  protected String getOzoneManagerServiceId(OzoneConfiguration configuration)
-      throws IOException {
-    Collection<String> omServiceIds;
-    String localOMServiceId = configuration.get(OZONE_OM_INTERNAL_SERVICE_ID);
-    if (localOMServiceId == null) {
-      LOG.info("{} is not defined, falling back to {} to find serviceID for "
-              + "OzoneManager if it is HA enabled cluster",
-          OZONE_OM_INTERNAL_SERVICE_ID, OZONE_OM_SERVICE_IDS_KEY);
-      omServiceIds = configuration.getTrimmedStringCollection(
-          OZONE_OM_SERVICE_IDS_KEY);
-    } else {
-      omServiceIds = Collections.singletonList(localOMServiceId);
-    }
-
-    String serviceId;
-    if (omServiceIds.isEmpty()) {
-      throw new IOException("No OzoneManager ServiceID configured to work "
-          + "with for Recon.");
-    } else {
-      serviceId = omServiceIds.iterator().next();
-      LOG.info("Using OzoneManager ServiceID '{}' for Recon", serviceId);
-    }
-    return serviceId;
   }
 
   @Provides
